Move some more classes to appropriate packages
