+++ app/src/main/java/com/crewly/duty/ryanair/RyanairDutyType.kt
@@ -1,4 +1,4 @@
-package com.crewly.duty
+package com.crewly.duty.ryanair
 
 /**
  * Created by Derek on 19/08/2018
