+++ app/src/main/java/com/crewly/utils/TestExtensions.kt
@@ -1,7 +1,7 @@
 package com.crewly.utils
 
 import com.crewly.duty.Duty
-import com.crewly.duty.RyanairDutyType
+import com.crewly.duty.ryanair.RyanairDutyType
 import com.crewly.duty.Sector
 import com.crewly.roster.RosterPeriod
 import com.crewly.roster.RosterPeriod.RosterDate
